backend/src/main/kotlin/ase/athlete_view/domain/user/service/TrainerService.kt
/*-
 * #%L
 * athlete_view
 * %%
 * Copyright (C) 2023 - 2024 TU Wien INSO ASE GROUP 5 WS2023
 * %%
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 * #L%
 */
package ase.athlete_view.domain.user.service

import ase.athlete_view.domain.user.pojo.dto.UserDTO
import ase.athlete_view.domain.user.pojo.entity.Trainer

interface TrainerService {
    /**
     * @return true if there is a trainer with a given code.
     */
    fun getByCode(code: String): Trainer?

    /**
     * resets the code of a trainer.
     */
    fun resetCode(user: UserDTO)

    /**
     * Accepts an athlete by a trainer.
     *
     * @param userDTO the trainer.
     * @param athleteDTO the athlete.
     */
    fun acceptAthlete(userDTO: UserDTO, id: Long)

    /**
     * Invites athletes by a trainer.
     *
     * @param id of the trainer.
     * @param emailList the list of emails of athletes.
     */
    fun inviteAthletes(id: Long, emailList: List<String>) {

    }
}
